refactored eventdatelabel to eventlabel, and moved it out of the folder with the index to be in only one file for readability
